ajout de la mise en page du site et fixe bug des commentaires

+++ backend/controllers/post.js
@@ -1,7 +1,7 @@
 const jwt = require('jsonwebtoken');
-
 const db = require("../models");
 const fs = require('fs');
+const xss = require('xss');
 
 exports.getAllPosts = (req, res) => {
     db.Post.findAll({
@@ -25,11 +25,13 @@ exports.createOnePost = (req, res) => {
     const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET');
     const id = decodedToken.userId;
 
+    const file = req.file ? `${req.protocol}://${req.get('host')}/images/${req.file.filename}` : null;
+
     const post = {
-        userId: id,
-        title: req.body.title,
-        content: req.body.content,
-        attachment: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+        userId: xss(id),
+        title: xss(req.body.title),
+        content: xss(req.body.content),
+        attachment: xss(file)
     };
 
     // création d'un post 
@@ -84,15 +86,21 @@ exports.deleteOnePost = (req, res) => {
     db.Post.findOne({
         where: {id: req.params.id || isAdmin == true}
     })
-    .then((post) => {    
-        const filename = post.attachment.split('/images/')[1];
-        fs.unlink(`images/${filename}`, () => {
+    .then((post) => {  
+        if (db.Post.attachment == null) {
             db.Comment.destroy({ where: { postId: req.params.id }})
             db.Post.destroy({ where: { id: req.params.id }})
             .then(() => res.status(200).json({ message: "Votre post a bien été supprimé" }))
             .catch((error) => res.status(400).json({ error }))  
-        })
-        
+        } else {
+            const filename = post.attachment.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                db.Comment.destroy({ where: { postId: req.params.id }})
+                db.Post.destroy({ where: { id: req.params.id }})
+                .then(() => res.status(200).json({ message: "Votre post a bien été supprimé" }))
+                .catch((error) => res.status(400).json({ error }))  
+            })
+        }  
     })
     .catch((error) => { 
         res.status(400).json({
